fix(api): send coordinates as numbers and expose phase_event_id on the wire
feat(dispatcher): artifact read path, date utils and geofence helpers

feat(dispatcher): per-phase evidence cards and resizable manifest panel

+++ backend/app/schemas/evidence.py
@@ -1,7 +1,6 @@
 """Pydantic v2 schemas for EvidenceArtifact."""
 
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 from typing import Optional
 
@@ -22,8 +21,8 @@ class EvidenceArtifactBase(BaseModel):
     captured_at: datetime
     captured_by_driver_id: Optional[UUID] = None
     captured_by_user_id: Optional[UUID] = None
-    captured_lat: Optional[Decimal] = None
-    captured_lng: Optional[Decimal] = None
+    captured_lat: Optional[float] = None
+    captured_lng: Optional[float] = None
 
 
 class EvidenceArtifactCreate(EvidenceArtifactBase):
@@ -34,8 +33,8 @@ class EvidenceArtifactUpdate(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
     # Artifacts are immutable after creation; only metadata fields may be patched.
-    captured_lat: Optional[Decimal] = None
-    captured_lng: Optional[Decimal] = None
+    captured_lat: Optional[float] = None
+    captured_lng: Optional[float] = None
 
 
 class EvidenceArtifactRead(EvidenceArtifactBase):
